refactor op call instruction (#8718)
* refactor op call instruction

* Add a blank line

* rm friend struct

* auto format by CI

* refine

* auto format by CI
