--- oneflow/core/eager/op_call_instruction_type.h
@@ -1,46 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef ONEFLOW_CORE_EAGER_OP_CALL_INSTRUCTION_TYPE_H_
-#define ONEFLOW_CORE_EAGER_OP_CALL_INSTRUCTION_TYPE_H_
-
-#include "oneflow/core/vm/instruction.h"
-#include "oneflow/core/vm/instruction_type.h"
-#include "oneflow/core/memory/memory_case.pb.h"
-
-namespace oneflow {
-namespace vm {
-
-class OpCallInstructionType final : public vm::InstructionType {
- public:
-  OpCallInstructionType() = default;
-  ~OpCallInstructionType() = default;
-
-  Maybe<void> Prepare(vm::Instruction* instruction) const override;
-  void Compute(vm::Instruction* instruction) const override;
-
-  InstructionFuseType fuse_type() const override { return kEnableInstructionFuseAtAnyPosition; }
-
-  std::string DebugName(const vm::Instruction& instruction) const override;
-
- protected:
- private:
-  Maybe<void> MaybeCompute(vm::Instruction* instruction) const;
-};
-
-}  // namespace vm
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_EAGER_OP_CALL_INSTRUCTION_TYPE_H_
